+++ src/FluxFlow.Components.Routing/RoutingErrorCodeNames.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Routing;
+
+public static class RoutingErrorCodeNames
+{
+    public const string OperationFailed = "routing.operation_failed";
+}